+++ src/main/java/showtime_corp/profile_vitaile/repository/ResumenRepository.java
@@ -2,12 +2,11 @@ package showtime_corp.profile_vitaile.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import showtime_corp.profile_vitaile.entity.Employability;
 import showtime_corp.profile_vitaile.entity.Resumen;
 
 import java.util.Optional;
 
 @Repository
-public interface ResumenRepository extends JpaRepository<Resumen, Integer> {
-    Optional<Resumen> findByUserId(Integer userId);
+public interface ResumenRepository extends JpaRepository<Resumen, Long> {
+    Optional<Resumen> findByUserId(Long userId);
 }
